Add support for dark/light themes
Uses the browser's preferred color scheme, rather than having a specific
toggle control. Could go for a toggle in the future but this seemed
simplest.

Loses syntax highlighting support. If I do end up posting code on here a
ton in the future I might figure out a way to add it back, but having
statically highlighted code embedded in the page looked awful (because
it couldn't switch depending on light/dark mode).